requirements.txt
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0
pytesseract==0.3.10 